@@ -27,13 +27,13 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "http-interop/http-factory-guzzle": "^1.0",
-        "hyperf/di": "^2.1",
-        "hyperf/framework": "^2.1",
-        "hyperf/guzzle": "^2.1",
-        "hyperf/scout": "^2.1",
-        "meilisearch/meilisearch-php": "^0.20.0"
+        "hyperf/di": "^2.2.0",
+        "hyperf/framework": "^2.2.0",
+        "hyperf/guzzle": "^2.2.0",
+        "hyperf/scout": "^2.2.0",
+        "meilisearch/meilisearch-php": "^0.26.1"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.14",
